\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary}
\alias{summary,Paths-method}

\title{Summery of the Paths Object}

\description{
	Shows the path count of the given \code{\link{Paths}} object.

\emph{Usage}

\code{summary(object)}
}

\arguments{
	\item{object}{Object of type \code{\link{Paths}}.}
}

\details{
	Given the \code{\link{Paths}} object, this method prints the path count that object contains.
}

\seealso{\code{\link{Paths}}, \code{\link{kBestShortestPaths}}}
\keyword{methods}
